Remove model dir and PSR-4 changes to fix Forge
