+++ app/Category.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Models;
+namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
-class DetailType extends Model
+class Category extends Model
 {
     public function group(): BelongsTo
     {
