+++ app/DetailType.php
@@ -1,17 +1,12 @@
 <?php
 
-namespace App\Models;
+namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
-class ServingSize extends Model
+class DetailType extends Model
 {
-    protected $fillable = [
-        'size_imperial',
-        'size_metric',
-    ];
-
     public function group(): BelongsTo
     {
         return $this->belongsTo(Group::class);
