+++ app/Http/Controllers/GroupController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Group;
+use App\Group;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\View\View;
@@ -16,7 +16,7 @@ class GroupController extends Controller
      */
     public function index()
     {
-        $groups = Group::all()->groupBy('category');
+        $groups = Group::all()->groupBy('category_id');
 
         return view('welcome')->with('groups', $groups);
     }
