+++ app/ServingSize.php
@@ -1,12 +1,17 @@
 <?php
 
-namespace App\Models;
+namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
-class Category extends Model
+class ServingSize extends Model
 {
+    protected $fillable = [
+        'size_imperial',
+        'size_metric',
+    ];
+
     public function group(): BelongsTo
     {
         return $this->belongsTo(Group::class);
